+++ packages/cli/src/commands/repo/fix.ts
@@ -22,11 +22,7 @@ import {
 } from '@backstage/cli-node';
 import { OptionValues } from 'commander';
 import fs from 'fs-extra';
-import {
-  resolve as resolvePath,
-  join as joinPath,
-  relative as relativePath,
-} from 'path';
+import { resolve as resolvePath, posix, relative as relativePath } from 'path';
 import { paths } from '../../lib/paths';
 
 /**
@@ -205,7 +201,7 @@ export function createRepositoryFieldFixer() {
   const rootDir = rootRepoField.directory || '';
 
   return (pkg: FixablePackage) => {
-    const expectedPath = joinPath(
+    const expectedPath = posix.join(
       rootDir,
       relativePath(paths.targetRoot, pkg.dir),
     );