+++ packages/integration-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/integration-react
 
+## 1.1.27-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/integration@1.11.0-next.0
+
 ## 1.1.26
 
 ### Patch Changes